refactor: EmailForm, emailAction
