+++ src/app/(auth)/forgot/password/(OTPForm)/OTPForm.tsx
@@ -1,0 +1,3 @@
+export default function OTPForm() {
+  return <div>Otp Form</div>
+}
